new linked script xresources

+++ .config/X11/link-correct-xresource.zsh
@@ -1,4 +1,8 @@
 #!/bin/zsh
 
 cd "$XDG_CONFIG_HOME/X11"
-ln -sf "Xresources-$HOST" "Xresources"
+if [[ -f "Xresources-$HOST" ]]; then
+    ln -sf "Xresources-$HOST" "Xresources"
+else
+    ln -sf "Xresources-desktop" "Xresources"
+fi
